+++ src/main/kotlin/cloud/glitchdev/rfu/config/categories/OtherSettings.kt
@@ -12,6 +12,11 @@ object OtherSettings : Category("Other") {
         }
     }
 
+    var emojis by boolean(true) {
+        name = Literal("Emojis")
+        description = Literal("Replaces triggers like :dog: with the respective high-res emoji.")
+    }
+
     var lobbyTracking by boolean(true) {
         name = Literal("Enable lobby tracking")
         description = Literal("Sends a message whenever you're in a lobby you've been in before.")