+++ src/main/kotlin/cloud/glitchdev/rfu/constants/Emoji.kt
@@ -1,0 +1,9 @@
+package cloud.glitchdev.rfu.constants
+
+object Emoji {
+    const val DOG_CHAR = "\uE001"
+
+    val ALL = mapOf(
+        ":dog:" to DOG_CHAR
+    )
+}