+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added hotspot sharing
 - Added party commands
 - Added /rfuscedit command to open the Sea Creature Edit window
+- Added Dog Emoji :dog: (I'm not addicted, okay?!)
 
 ### Changes: 
 - Made the report/delete button in rfupf last a little bit so it is easier to click